+++ bench/scenario/load.go
@@ -202,7 +202,7 @@ scenarioLoop:
 			//定期的にconditionを見に行くシナリオ
 			request := service.GetIsuConditionRequest{
 				StartTime:        nil,
-				CursorEndTime:    uint64(dataExistTimestamp),
+				CursorEndTime:    dataExistTimestamp,
 				CursorJIAIsuUUID: "",
 				ConditionLevel:   "info,warning,critical",
 				Limit:            nil,
@@ -233,7 +233,7 @@ scenarioLoop:
 				var conditionsTmp []*service.GetIsuConditionResponse
 				request = service.GetIsuConditionRequest{
 					StartTime:        nil,
-					CursorEndTime:    uint64(conditions[len(conditions)-1].Timestamp),
+					CursorEndTime:    conditions[len(conditions)-1].Timestamp,
 					CursorJIAIsuUUID: "",
 					ConditionLevel:   "info,warning,critical",
 					Limit:            nil,
@@ -262,7 +262,7 @@ scenarioLoop:
 			if solvedCondition != model.IsuStateChangeNone && lastSolvedTime[targetIsu.JIAIsuUUID].Before(time.Unix(findTimestamp, 0)) {
 				//graphを見る
 				virtualDay := (findTimestamp / (24 * 60 * 60)) * (24 * 60 * 60)
-				_, _, errs := browserGetIsuGraphAction(ctx, user.Agent, targetIsu.JIAIsuUUID, uint64(virtualDay),
+				_, _, errs := browserGetIsuGraphAction(ctx, user.Agent, targetIsu.JIAIsuUUID, virtualDay,
 					func(res *http.Response, graph []*service.GraphResponse) []error {
 						return []error{} //TODO: 検証
 					},
@@ -285,7 +285,7 @@ scenarioLoop:
 
 			//TODO: graphを見に行くシナリオ
 			virtualToday := (dataExistTimestamp / (24 * 60 * 60)) * (24 * 60 * 60)
-			_, graphToday, errs := browserGetIsuGraphAction(ctx, user.Agent, targetIsu.JIAIsuUUID, uint64(virtualToday),
+			_, graphToday, errs := browserGetIsuGraphAction(ctx, user.Agent, targetIsu.JIAIsuUUID, virtualToday,
 				func(res *http.Response, graph []*service.GraphResponse) []error {
 					//検証前にデータ取得
 					user.GetConditionFromChan(ctx)
@@ -301,7 +301,7 @@ scenarioLoop:
 			}
 
 			//前日のグラフ
-			_, _, errs = browserGetIsuGraphAction(ctx, user.Agent, targetIsu.JIAIsuUUID, uint64(virtualToday-60*60),
+			_, _, errs = browserGetIsuGraphAction(ctx, user.Agent, targetIsu.JIAIsuUUID, virtualToday-60*60,
 				func(res *http.Response, graph []*service.GraphResponse) []error {
 					return []error{} //TODO: 検証
 				},
@@ -324,7 +324,7 @@ scenarioLoop:
 
 			//悪いものがあれば、そのconditionを取る
 			if errorEndAtUnix != 0 {
-				startTime := uint64(errorEndAtUnix - 60*60)
+				startTime := errorEndAtUnix - 60*60
 				//MEMO: 本来は必ず1時間幅だが、検証のためにdataExistTimestampで抑える
 				cursorEndTime := errorEndAtUnix
 				if dataExistTimestamp < cursorEndTime {
@@ -332,7 +332,7 @@ scenarioLoop:
 				}
 				request := service.GetIsuConditionRequest{
 					StartTime:        &startTime,
-					CursorEndTime:    uint64(cursorEndTime),
+					CursorEndTime:    cursorEndTime,
 					CursorJIAIsuUUID: "",
 					ConditionLevel:   "warning,critical",
 					Limit:            nil,
@@ -508,7 +508,7 @@ scenarioLoop:
 			//定期的にconditionを見に行くシナリオ
 			request := service.GetIsuConditionRequest{
 				StartTime:        nil,
-				CursorEndTime:    uint64(dataExistTimestamp),
+				CursorEndTime:    dataExistTimestamp,
 				CursorJIAIsuUUID: "z",
 				ConditionLevel:   "warning,critical",
 				Limit:            nil,
@@ -539,7 +539,7 @@ scenarioLoop:
 				var conditionsTmp []*service.GetIsuConditionResponse
 				request = service.GetIsuConditionRequest{
 					StartTime:        nil,
-					CursorEndTime:    uint64(conditions[len(conditions)-1].Timestamp),
+					CursorEndTime:    conditions[len(conditions)-1].Timestamp,
 					CursorJIAIsuUUID: conditions[len(conditions)-1].JIAIsuUUID,
 					ConditionLevel:   "warning,critical",
 					Limit:            nil,
@@ -567,7 +567,7 @@ scenarioLoop:
 				if solvedConditions[targetIsuID] != model.IsuStateChangeNone && lastSolvedTime[targetIsuID].Before(time.Unix(timestamp, 0)) {
 					//graphを見る
 					virtualDay := (timestamp / (24 * 60 * 60)) * (24 * 60 * 60)
-					_, _, errs := browserGetIsuGraphAction(ctx, user.Agent, targetIsuID, uint64(virtualDay),
+					_, _, errs := browserGetIsuGraphAction(ctx, user.Agent, targetIsuID, virtualDay,
 						func(res *http.Response, graph []*service.GraphResponse) []error {
 							return []error{} //TODO: 検証
 						},
